Add files via upload
This program demonstrate Caesar cipher encryption and decryption of a file. The program asks the user for filename (extension ) and Output of encrypted file. It also ask the operation to perform and if it encrypted u will see it labelled as ENCRYPTED at top of the file. And once done you can exit from program.