Create caesarcipher for key more than one digit
This takes multi-digit key value, previous one was constrained to take only 0-25 as there are only 26 alphabets in the English language. So it couldn't process more than 25 as it becomes invalid shifts and making it inconsistent, but this multi-digit key input there is no key constraint(but key have to integer) don't give strings or float value as key it will crash or result is inconsistent.